-END()
Under eager evaluation, the END() command will lead to unexpected
behavior. The BREAK() command has the same problem and is yet to be
solved.

It is not certain wheher it is necessary to switch to a lazy approach
which is not very safe as multiple scripts and commands run concurrently